Create README.md
I was given an assignment to:
Write a program that asks the user to enter a salary.  If salary is more than 100000  then output "invalid". If salary is less than 10000 output "invalid".  If salary is less than 30000  (but more than 10000)  give a 20% raise. If salary is less than 40000 give a 10% raise. If salary is less than 50000  give a 5% raise. Output the new salary.

just want to make sure im on the right track here